refactor: Split Contact component to another files
